Created service URL helper
